refactor reusable components into other screens and finish screen for PlayerSelectionScreen
