+++ src/Screens/HomeScreen/HomeScreen.css
@@ -1,42 +1,15 @@
-/* Main HomeScreen */
-.screen {
-    height: 100%;
-}
-
-.top {
-    width: 100%;
-    height: 50%;
-    background-color: #979797;
-    z-index: 0;
-}
-
-/* Bottom Half */
-.bottom {
-    width: 100%;
-    height: 50%;
-    z-index: 1;
-    background-color: white;
-    display: grid;
-    grid-template-columns: 15% auto 15%;
-    grid-template-rows: 5fr 1fr;
-    position: relative;
-}
-
-.inner {
-    grid-column: 2/3; 
-    display: grid;
-    grid-template-rows: 1fr 1fr 1fr 1fr 1fr;
-    grid-row-gap: 15px;
-    height: inherit;
-}
-
 .title {
     text-align: center;
+    grid-row: 1;
 }
 
 /* Or Component */
 .orDiv {
-    width: 100%; height: 20px; border-bottom: 1px solid black; text-align: center;
+    width: 80%;
+    height: 20px; 
+    border-bottom: 1px solid black; 
+    text-align: center;
+    margin: 20px auto 20px auto;
 }
 
 .orText {
