+++ src/Screens/HomeScreen/HomeScreen.jsx
@@ -2,7 +2,10 @@ import React from 'react'
 import "./HomeScreen.css"
 
 import Button from "../../SharedComponents/Button/Button"
-import { TopHome } from './TopHome';
+import Top from "../../SharedComponents/Top/Top"
+import Card from "../../SharedComponents/Card/Card"
+import Screen from "../../SharedComponents/Screen/Screen"
+import Bottom from "../../SharedComponents/Bottom/Bottom"
 
 function Or() {
   return (
@@ -14,30 +17,24 @@ function Or() {
   );
 }
 
-function BottomHome() {
-  return (
-    <div className="bottom">
-      <div className="inner">
-        <div className="title">
-          <h1>A party game for horrible people</h1>
-        </div>
-        <Button text="Join Party"/>
-        <Or />
-        <Button text="Create Game"/>
-      <div className="footer">
-        Like and share us on <a href="http://www.facebook.com/yusufameri"> Facebook!</a>
-      </div>
-      </div>
-    </div>
-  )
-}
-
 function HomeScreen() {
   return (
-    <div className="screen">
-      <TopHome />
-      <BottomHome />
-    </div>
+    <Screen>
+      <Top>
+        <Card cardType="Title" text="Cards Against Humanity" />
+      </Top>
+      <Bottom>
+        <div className="title">
+            <h1>A party game for horrible people</h1>
+          </div>
+          <Button text="Join Party"/>
+          <Or />
+          <Button text="Create Game"/>
+          <div className="footer">
+            Like us on <a href="http://www.facebook.com/yusufameri"> Facebook!</a>
+          </div>
+      </Bottom>
+    </Screen>
   );
 }
