--- src/Screens/HomeScreen/TopHome.jsx
@@ -1,10 +1,0 @@
-import Card from '../../SharedComponents/Card/Card';
-import React from 'react'
-
-export function TopHome() {
-  return (
-    <div className="top">
-      <Card cardType="Title" text="Cards Against Humanity" />
-    </div>
-  );
-}
