+++ src/Screens/PlayerSelectionScreen/PlayerSelectionScreen.jsx
@@ -1,25 +1,18 @@
 import React from 'react'
 import "./PlayerSelectionScreen.css"
-import Card from "../../SharedComponents/Card/Card"
-import Bottom from "../../SharedComponents/Bottom/Bottom"
+
 import Screen from "../../SharedComponents/Screen/Screen"
 import Top from "../../SharedComponents/Top/Top"
 import HeaderMenu from "../../SharedComponents/HeaderMenu/HeaderMenu"
+import DropCardSpace from "../../SharedComponents/DropCardSpace/DropCardSpace"
+import Bottom from "../../SharedComponents/Bottom/Bottom"
+import CardCarousel from "../../SharedComponents/CardCarousel/CardCarousel"
 
 // Black Card
 let QCard = {
   cardType: "Q",
   text: "Some Clever statement or question",
   id: 0
-}
-
-function DropCardSpace(props) {
-  return (
-    <div className="drop-space">
-      <Card cardType= "placeholder"/>
-      <Card {...QCard}/>
-    </div>
-  );
 }
 
 // Mocking cards data to demo usage in props...
@@ -41,22 +34,12 @@ let cards = [
   }
 ];
 
-function CardCarousel(props) {
-  return (
-    <div className="scrolling-wrapper">
-      {
-        props.cards.map((card) => <Card text={card.text}/>)
-      }
-    </div>
-  );
-}
-
 function PlayerSelectionScreen() {
   return (
     <Screen>
       <Top>
         <HeaderMenu text="Yusuf is the Judge" timeLeft={49}/>
-        <DropCardSpace qcard = {QCard}/>
+        <DropCardSpace QCard = {QCard}/>
       </Top>
       <Bottom>
         <CardCarousel cards ={cards}/>
