+++ src/SharedComponents/Button/Button.css
@@ -11,4 +11,5 @@
     cursor: pointer;
     margin-left: auto;
     margin-right: auto;
+    margin-bottom: 20px;
 }
