+++ src/SharedComponents/CardCarousel/CardCarousel.css
@@ -1,0 +1,11 @@
+.scrolling-wrapper {
+    display: flex;
+    flex-wrap: nowrap;
+    overflow-x: auto;
+}
+
+.card {
+  flex: 0 0 auto;
+  margin-right: 11px;
+}
+
