--- src/SharedComponents/CardList/CardList.css
@@ -1,5 +1,0 @@
-.cardList > .card {
-    display: inline-flex;
-    margin-right: 20px;
-    margin-bottom: 20px;
-}
