+++ src/SharedComponents/DropCardSpace/DropCardSpace.css
@@ -1,0 +1,8 @@
+.drop-space {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: 1fr;
+  padding-left: 10px;
+  padding-right: 10px;
+  padding-top: 20px;
+}
