+++ src/index.js
@@ -3,6 +3,16 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import './normalize.css'
 
-import PlayerSelectionScreen from './Screens/PlayerSelectionScreen/PlayerSelectionScreen'
+// import PlayerSelectionScreen from './Screens/PlayerSelectionScreen/PlayerSelectionScreen'
+// ReactDOM.render(<PlayerSelectionScreen />, document.getElementById('root'));
 
-ReactDOM.render(<PlayerSelectionScreen />, document.getElementById('root'));
+// import HomeScreen from "./Screens/HomeScreen/HomeScreen"
+// ReactDOM.render(<HomeScreen />, document.getElementById('root'));
+
+// import JoinPartyScreen from "./Screens/JoinPartyScreen/JoinPartyScreen"
+// ReactDOM.render(<JoinPartyScreen />, document.getElementById('root'));
+
+import StartGameScreen from "./Screens/StartGameScreen/StartGameScreen"
+ReactDOM.render(<StartGameScreen />, document.getElementById('root'));
+
+
